+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/IntegrationTests.cs
@@ -24,23 +24,29 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 {
     public class IntegrationTests
     {
-        private const string CollectorEndpointEnvVarName = "OTEL_EXPORTER_OTLP_ENDPOINT";
-        private static readonly string CollectorEndpoint = SkipUnlessEnvVarFoundFactAttribute.GetEnvironmentVariable(CollectorEndpointEnvVarName);
+        private const string CollectorHostnameEnvVarName = "OTEL_COLLECTOR_HOSTNAME";
+        private static readonly string CollectorHostname = SkipUnlessEnvVarFoundTheoryAttribute.GetEnvironmentVariable(CollectorHostnameEnvVarName);
 
+        [InlineData(OtlpExportProtocol.Grpc, ":4317")]
+        [InlineData(OtlpExportProtocol.HttpProtobuf, ":4318/v1/traces")]
         [Trait("CategoryName", "CollectorIntegrationTests")]
-        [SkipUnlessEnvVarFoundFact(CollectorEndpointEnvVarName)]
-        public void ExportResultIsSuccess()
+        [SkipUnlessEnvVarFoundTheory(CollectorHostnameEnvVarName)]
+        public void ExportResultIsSuccess(OtlpExportProtocol protocol, string endpoint)
         {
 #if NETCOREAPP3_1
             // Adding the OtlpExporter creates a GrpcChannel.
-            // This switch must be set before creating a GrpcChannel/HttpClient when calling an insecure gRPC service.
+            // This switch must be set before creating a GrpcChannel when calling an insecure HTTP/2 endpoint.
             // See: https://docs.microsoft.com/aspnet/core/grpc/troubleshoot#call-insecure-grpc-services-with-net-core-client
-            AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            if (protocol == OtlpExportProtocol.Grpc)
+            {
+                AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            }
 #endif
 
             var exporterOptions = new OtlpExporterOptions
             {
-                Endpoint = new System.Uri($"http://{CollectorEndpoint}"),
+                Endpoint = new System.Uri($"http://{CollectorHostname}{endpoint}"),
+                Protocol = protocol,
             };
 
             var otlpExporter = new OtlpTraceExporter(exporterOptions);
@@ -56,7 +62,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             using var tracerProvider = builder.Build();
 
             var source = new ActivitySource(activitySourceName);
-            var activity = source.StartActivity("Test Activity");
+            var activity = source.StartActivity($"{protocol} Test Activity");
             activity?.Stop();
 
             Assert.Single(delegatingExporter.ExportResults);