+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -2,12 +2,24 @@
 
 ## Unreleased
 
+## 1.2.0-rc1
+
+Released 2021-Nov-29
+
 * Added configuration options for `MetricReaderType` to allow for configuring
   the `OtlpMetricExporter` to export either manually or periodically.
   ([#2674](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2674))
 * Added validation that unsecure channel is configured correctly when using
   .NET Core 3.1 for gRPC-based exporting.
   ([#2691](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2691))
+
+* The internal log message used when OTLP export client connection failure occurs,
+  will now include the endpoint uri as well.
+  ([#2686](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2686))
+
+* Support `HttpProtobuf` protocol with metrics & added `HttpClientFactory`
+  option
+  ([#2696](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2696))
 
 ## 1.2.0-beta2